# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Medicine": {
            "get": {
                "description": "list Medicine entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List Medicine entities",
                "operationId": "list-Medicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Medicine"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Medicine",
                "operationId": "create-Medicine",
                "parameters": [
                    {
                        "description": "Medicine entity",
                        "name": "Medicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Medicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Medicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/dispensemedicines": {
            "get": {
                "description": "list dispensemedicine entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List dispensemedicine entities",
                "operationId": "list-dispensemedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.DispenseMedicine"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create dispensemedicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create dispensemedicine",
                "operationId": "create-dispensemedicine",
                "parameters": [
                    {
                        "description": "DispenseMedicine entity",
                        "name": "dispensemedicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DispenseMedicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.DispenseMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/dispensemedicines/{id}": {
            "get": {
                "description": "get dispensemedicine by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a dispensemedicine entity by ID",
                "operationId": "get-dispensemedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DispenseMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.DispenseMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update dispensemedicine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a dispensemedicine entity by ID",
                "operationId": "update-dispensemedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DispenseMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "DispenseMedicine entity",
                        "name": "dispensemedicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.DispenseMedicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.DispenseMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get dispensemedicine by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a dispensemedicine entity by ID",
                "operationId": "delete-dispensemedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DispenseMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.DispenseMedicine": {
            "type": "object",
            "properties": {
                "annotation": {
                    "type": "integer"
                },
                "datetime": {
                    "type": "string"
                },
                "pharmacist": {
                    "type": "integer"
                },
                "prescription": {
                    "type": "integer"
                }
            }
        },
        "controllers.Medicine": {
            "type": "object",
            "properties": {
                "levelOfDangerousID": {
                    "type": "integer"
                },
                "medicineTypeID": {
                    "type": "integer"
                },
                "unitOfMedicineID": {
                    "type": "integer"
                }
            }
        },
        "ent.Annotation": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AnnotationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AnnotationEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "messages": {
                    "description": "Messages holds the value of the \"messages\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Bill": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "annotation": {
                    "description": "Annotation holds the value of the \"annotation\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.DispenseMedicine": {
            "type": "object",
            "properties": {
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DispenseMedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DispenseMedicineEdges": {
            "type": "object",
            "properties": {
                "annotation": {
                    "description": "Annotation holds the value of the annotation edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Annotation"
                },
                "bills": {
                    "description": "Bills holds the value of the Bills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bill"
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                },
                "prescription": {
                    "description": "Prescription holds the value of the prescription edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prescription"
                }
            }
        },
...
        "ent.DrugAllergy": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "DateTime holds the value of the \"dateTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DrugAllergyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DrugAllergyEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.LevelOfDangerous": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LevelOfDangerousQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LevelOfDangerousEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.LevelOfDangerousEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the Medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                }
            }
        },
        "ent.Medicine": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "brand": {
                    "description": "Brand holds the value of the \"brand\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineEdges"
                },
                "howtouse": {
                    "description": "Howtouse holds the value of the \"howtouse\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "serial": {
                    "description": "Serial holds the value of the \"serial\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MedicineEdges": {
            "type": "object",
            "properties": {
                "drugallergys": {
                    "description": "Drugallergys holds the value of the drugallergys edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DrugAllergy"
                    }
                },
                "levelOfDangerous": {
                    "description": "LevelOfDangerous holds the value of the LevelOfDangerous edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LevelOfDangerous"
                },
                "medicineType": {
                    "description": "MedicineType holds the value of the MedicineType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineType"
                },
                "medicinepresciption": {
                    "description": "Medicinepresciption holds the value of the medicinepresciption edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                },
                "order": {
                    "description": "Order holds the value of the order edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                },
                "unitOfMedicine": {
                    "description": "UnitOfMedicine holds the value of the UnitOfMedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnitOfMedicine"
                }
            }
        },
        "ent.MedicineType": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OrderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OrderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "total": {
                    "description": "Total holds the value of the \"total\" field.",
                    "type": "number"
                }
            }
        },
...
        "ent.Pharmacist": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PharmacistQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PharmacistEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Prescription": {
            "type": "object",
            "properties": {
                "Value": {
                    "description": "Value holds the value of the \"Value\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrescriptionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrescriptionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.UnitOfMedicine": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UnitOfMedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnitOfMedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}